feat: made sliceArray into util fucntion
