+++ packages/vendure-plugin-popularity-scores/src/popularity-scores.service.ts
@@ -15,6 +15,7 @@ import {
 } from '@vendure/core';
 import { PLUGIN_INIT_OPTIONS, loggerCtx } from './constants';
 import { PopularityScoresPluginConfig } from './popularity-scores.plugin';
+import { sliceArray } from './utils';
 @Injectable()
 export class PopularityScoresService implements OnModuleInit {
   private jobQueue!: JobQueue<{
@@ -143,17 +144,6 @@ export class PopularityScoresService implements OnModuleInit {
 
       const uniqueProductIds = [...new Set(productIds)];
       if (uniqueProductIds.length) {
-        const sliceArray = (arr: number[], chunkSize: number): number[][] => {
-          const result = [];
-          for (let i = 0; i < arr.length; i += chunkSize) {
-            if (i + chunkSize < arr.length) {
-              result.push(arr.slice(i, i + chunkSize));
-            } else {
-              result.push(arr.slice(i));
-            }
-          }
-          return result;
-        };
         let score = 0;
         const chunkedProductIds = sliceArray(
           uniqueProductIds,
